requirements.txt
fastapi
uvicorn
jinja2
fitz
python-multipart
PyMuPDF
openai
google-generativeai
pdf2image
pytesseract
